chickenRun
This program is a racing simulator game, where the user places bets -
as one of the bettors - on one of four chickens in a race.

Full details in readMe.

A WIP, still to improve; allow user to reset game; Update each bettors
wallet amount following the race, so mutliple races can be run in
succession.